Help with context
Thanks for taking a look! The issue I am running into is that in my schema.ex, the subscription is trying to use the context. Everything compiles just fine, but if you look at the media folder that is also in this repo, you can see that while everything still runs, I am getting a runtime error, and my subscription messages don't pop up in the subscription window while I am using the context. from what i can tell, the main issue is that it doesn't resolve properly, thus the notification does not send. Taking out the context reference to person will let it work fine.